requirements.txt
amqp==5.3.1
beautifulsoup4==4.13.4
billiard==4.2.1
blinker==1.9.0
celery==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.7.0
Flask==3.1.0
flask-cors==5.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.5.3
MarkupSafe==3.0.2
prompt_toolkit==3.0.51
python-dateutil==2.9.0.post0
redis==5.0.3
requests==2.32.3
six==1.17.0
soupsieve==2.7
Sublist3r==1.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
